Deal with connection errors
If the HTTPUrlConnection throws certain exceptions, for some reason it would
cause the async task's postExecute method never to be called, and some progress
indicators to never disappear.

Also, set a read timeout -- if we don't have this, then connections that don't
complete for some reasons block indefinitely. We want them to stop so the user
can restart them (and hopefully they will work).