New ObaApi
This is the start of a replacement for the ObaApi.
This new interface better handles the new V2 REST API and allows
for less custom deserialization. It also accepts that certain
elements have variable content models, so each response has to
be tied to its specific request.

When this is complete, the old API will be deprecated,
all operations in the busbot app will be replaced, and the old API
eventually deleted.