New trip service
This is a complete refactor of the trip service.
The impetus mostly was the NPEs we're getting on API 5+ because of
the new way services are controlled, and adjusting the code as
necessary. However, this new setup makes it a bit more difficult
to do what we were doing with the service.

This is just as well, since the original code had problems and
could easily get into a state that drained the users battery.
The new code eliminates this by setting an alarm rather than
sleeping with a wake lock. There are no longer wake locks in the
service, and so there's no way we can drain the battery like
we did.

It also uses TripAlerts stored in the database to ensure that
when an alert is cancelled, we know it's cancelled rather than
using scheduling acrobatics to try not to reschedule a cancelled
alert.

Overall, much better code.