Switch back to non-fragment MapActivity
Since we can't use fragments and the map activity, MapFragment turns
into an Activity base class. We also can't use the LoaderManager, but
fortunately we can use a Loader so that helps.

There may still need to be some cleanup/management of loaders that we
need to do since we're no longer using the loader manager, but for now
it works.