Begin allowing for mock connections
I'd like to begin using canned responses for the unit tests, but in
order to do that I need a way of mocking the responses. Luckily the
ObaConnection stuff put in place makes it slightly easier to do that,
and this allows for connections to be completely replaced by another
mock class.

Implementation to come.