Big fixes / improvements to mock
Fix the reason why the uri didn't look like a real URI -- we should
have just been setting the path rather than appending it.

Replace the old refresh_test_responses script with a new script that
gets the URLs and destination files from the urimap, eliminating the
need to specify the URLs and paths twice (DRY). Because we can't really
parse JSON in bash, make it python.