Switch to new Regions API
This removes the code to access Google Docs, and with it all the crud
that we needed to do to get cookies and redirects to work. Woohoo!

Since this API can be accessed similiary to the other OBA APIs, move
the objects into the oba.elements package so they can be filled in
through JSON.