Refactor GeoPoint into maps package
GeoPoint is part of Android Maps API v1 library (i.e., Google APIs), which has been deprecated and is no longer supported.  We plan to move to Maps API v2 and the new Location Services platform shortly, which are both part of Google Play Services.  Neither of these services include the GeoPoint class, so we need to isolate GeoPoint into the maps package (and replace it with normal Android platform Location object in the remainder of the app) so it can be easily extracted when we move to the new APIs and remove the Google APIs library.  Additionally, this also enables the core OBA code to be used on platforms that don't support maps in general (or the Google APIs library), such as Google Glass.