Refresh the favorite state / custom stop name in sliding panel header
This fixes a bug where the star didn't refresh on sliding panel - If you selected a bus stop on the map and added a star to it, Go to menu > Starred stops, Select the starred stop, Un-star it, Navigate back to the map, the bus stop was still marked with the star.  This patch ensures that the favorite state and custom stop names are refreshed in onResume() of the fragment, so when returning to the fragment from another Activity the correct information is displayed.