Use fully qualified reference to AmazonMap/GoogleMap to avoid import statements
When filter copying the Maps API v2 classes from the Google to Amazon build flavor (#254), the filtering of the import statement for "com.google.android.gms.maps.GoogleMap" results in "com.amazon.geo.mapsv2.AmazonMap".  This in itself is fine, but the problem is that when "Optimize imports" is executed in Android Studio when committing changes to the maps classes, the ordering of the imports changes when they are sorted in alphabetical order.  This results in an infinite loop (with part of the process being executed manually) where:

1. the Google build flavor classes are copied to the Amazon build flavor, so they match identically (with the exception of the filtered strings)
2. The change is committed, but "optimize imports" changes the order of the import statements just before commit (and therefore the committed files don't identically match)
3. When the source is rebuilt, Gradle detects that the files don't match (because Android Studio optimize imports altered the import order) and executes 1) above again, which results in another change that must be committed via version control

This patch avoids this scenario entirely by referencing the GoogleMap and AmazonMap objects by their fully qualified class names, which does not require the import statement.