Show MapFragment when starting HomeActivity with an Intent that is related to the map
Previously, the HomeActivity NavigationDrawerFragment always showed the last selected fragment.  This is incorrect behavior when HomeActivity is started with an Intent from another Activity that is designed to show something on a map.  Previous the following behavior occurred:

* Tap on Favorite Stop in the "Starred Stops" fragment.  This showed the arrivals list.  Tap on the Map icon in the ActionBar to show the location of the stop on the map.  This restarts HomeActivity, but since the "Starred Stops" fragment is the most recently used fragment, it shows this, instead of the stop location on the map.
* Tap on Favorite Stop in the "Starred Stops" fragment.  This showed the arrivals list.   Tap on an arrival time and select "show route information".  Then tap on Map icon in ActionBar to show the route on the map.  This restarts HomeActivity, but since the "Starred Stops" fragment is the most recently used fragment, it shows this, instead of the stop location on the map.

This patch checks for the presence of the Intent extras bundle that indicates the HomeActivity is being started to show something on the map, and if so, switches to show the MapFragment.