Fix "show all" clickable span in header
There was an issue with the "show all" clickable span in the header not working - you'd tap on "show all" to remove the existing route filter, and it would have no effect on the route filter (when the header was in the sliding panel) - it would still only show the filtered routes.  It turns out that because we're recycling views, the ClickableSpan from the *previously selected* ArrivalsListFragment (from a previous tap on the map) still existed and was swallowing the tap event, and setting the filter on the previously selected fragment and stop_id.

This patch cleans up some of the initialization code surrounding the show_all TextView, and most importantly removes any previous ClickableSpans from the TextView before setting the new ClickableSpan for the currently selected stop_id/ArrivalsListFragment.