Change to setting scrollable view for SlidingUpPanel programmatically.
`sothree:umanoScrollableView` must point directly to the ListView, and we don't have access to this in the XML because we're programmatically creating the ListFragment.  We're now setting this via a callback from the ArrivalListFragment.onResume() (already in place) to the HomeActivity.  We can't directly set the ListView when creating the ArrivalListFragment in HomeActivity because the ListView doesn't yet exist at that point - hence the callback.