Only force reset of HomeActivity if the custom OTP URL actually changes
Before this patch, every time you tap on Preferences back arrow in Action Bar it resets HomeActivity.  This issue was triggered by 8f2a86e4be33e9ef38c617072df270222144eb53, which ensures that the preference summary is always up to date with the OTP Custom API URL.  Checking if the preference changed should be done in onPreferenceChange(), not checkPreferenceSummary(), which is what this patch changes.