Bump versionCode
We burned versionCode 80 - as after we uploaded and submitted the APK Google rejected it with the following vulnerability message:

------
onReceivedSslError

Detected in APK 80
Your app is using an unsafe implementation of WebViewClient.onReceivedSslError handler. Please see this Google Help Center article for details, including the deadline for fixing the vulnerability.
Vulnerable classes:
•	com.microsoft.live.AuthorizationRequest$OAuthDialog$AuthorizationWebViewClient
------

And we can't upload another APK with versionCode 80, so we'll bump to 81.