Override ES SDK social comment icon with PNG instead of vector
As discussed in https://github.com/OneBusAway/onebusaway-android/pull/835, it seems that an XML vector drawable doesn't properly override a library PNG drawable with the same name.  So, we'll switch to using PNGs for this in the application.  Using this method, Android Studio still shows the library icon when looking at the layouts in OBA Android, but when running the app the correct PNG drawables from the OBA app are shown.