Use android:src instead of app:srcCompat for resource drawables
During the course of PR #835 (commit https://github.com/OneBusAway/onebusaway-android/commit/1ca2f1c23263db2bc579b05d243aefac013a177b), these were changed to use the vector drawables bundled in the project.  However, we decided to remove these and rely on the using the PNGs drawables bundled with the project instead, as this matched (and will override) the PNG drawables in the EmbeddedSocial SDK.  So, we're no longer using vector drawables for these resources, and therefore can remove the srcCompat reference.