Fix logic when getting new location and triggering NavProvider
The logic here was obviously wrong - "if (mLastLocation != null) {" was called right after "if (mLastLocation == null) {", which never would have executed.  I believe it only worked because there were two bugs - locationUpdated was called again inside the logging IF statement, which was always on.  As a result the nav provider was still always triggered by the new location.  Also, the location always got passed into the NavProvider if it was a duplicate (this should have been if NOT a duplicate, I assume???)

This patch fixes (or simplifies, at least) the logic to always trigger the NavProvider:
1. If it's the first location
2. If it's a new location that is not a duplicate of the previous location (based on time, lat, and long)

Logging is moved after this IF block so it occurs in all cases if it's configured

I've also added a unit test for LocationUtils.isDuplicate().

This needs more testing in the field to confirm that the LocationService is properly triggered the navProvider.