Fix #953 - Call ES init prior to the initialization of ES Fragments
This is a workaround to make sure ES SDK is initialized in case we startup to ES Fragments.  The previous workaround that listens for an onStart() event from an application-wide ProcessLifecycleOwner Observer doesn't happen earlier enough in the Fragment lifecycle and results in an NPE.  This patches ensures that when starting HomeActivity, which could start up directly to showing an ES fragment, we call ES init before initializing the fragment being shown to the user.

When ES SDK no longer runs servers in the background Application.setUpSocial() can be made private again and all ES SDK initialization can happen in Application.onCreate()