fix: Initial work towards location permission prompt
The "No thanks" option for prompting location was previously broken (we kept prompting the user for location permissions), so fix that too.

Still prompts twice for permissions on start up due to multiple callbacks (needs to be fixed), but other than that it improves on previous iteration. Need to test cases when user allows permission too.