fix: Isolate physical activity and background location permission requests
When targeting Android 11, these must be requested separately otherwise neither is granted. This requires chaining the permission requests based on the granted callback of the first. The only way I could get this not to interfere with the existing permissions requested in BaseMapFragment is to use ActivityResultLauncher, which requires the two new androidx library dependencies (for Fragments and Activities, respectively). ActivityResultLauncher must be set up in HomeActivity.onCreate(), which required some refactoring to trigger the permission request when the user opts into the study in TravelBehaviorManager.