chore: Target SDK version 31 (Android 12) (#1087)
* chore: Initial work towards targeting SDK version 31

Currently this compiles but fails at runtime due to lack of FLAG_IMMUTABLE or FLAG_MUTABLE on PendingIntents

* chore: Add mutability to pending intents

Since these are all related to alarms, reply in notifications, or location/activity requests, they should all be mutable per https://developer.android.com/guide/components/intents-filters#CreateImmutablePendingIntents

Also bump Work Manager and Firebase dependencies as these were causing the same failures.

* ci: Update to latest setup instructions

From https://github.com/ReactiveCircus/android-emulator-runner

* ci: Use Java 11 again

* chore: Add explicit Gson dependency

Maybe previously this was being drawn in by another dependency that has since removed it?

* ci: Try just the google_apis

* chore: Ensure COARSE location permission is requested when FINE is

This also changes most of the app logic (with the exception of the opt-in travel behavior data collection, which still requires FINE) to allow EITHER COARSE or FINE permissions, so for example the coarse location will be shown on the map if the user has chosen COARSE location (but not FINE).

* fix: Isolate physical activity and background location permission requests

When targeting Android 11, these must be requested separately otherwise neither is granted. This requires chaining the permission requests based on the granted callback of the first. The only way I could get this not to interfere with the existing permissions requested in BaseMapFragment is to use ActivityResultLauncher, which requires the two new androidx library dependencies (for Fragments and Activities, respectively). ActivityResultLauncher must be set up in HomeActivity.onCreate(), which required some refactoring to trigger the permission request when the user opts into the study in TravelBehaviorManager.

* chore: Try to fix backup/restore with scoped storage

Restore is still broken.

* fix: Fix restore by presenting user with file picker to backup file

* fix: Add new permission that's now required for alarms

Also add note as to why we're not using setAlarmClock()

* ci: Try removing Java installation

* chore: Bump AGP and Gradle

* ci: Add Java 11 back