chore: Better handling of bad arrival/departure times
Now we check for the presence of valid (i.e., positive integer) predicted arrival or departure information from JSON before setting the boolean saying if predicted info is included.

Adds unit tests to check if prediction information is included or not, and to handle the case of bad data where the boolean in the JSON is set to true but the JSON doesn't contain valid predicted arrival/departure times.

Closes #1083