Fixes the trip planner
The trip planner had stopped working due to some overzealous ProGuard rules. This fix just blindly `-keep`s all of the `org.geojson` package. It would be great to be more judicious in the future, but for now I just want to get this feature working for our end users again. (Ostensibly there was a bump of the GeoJSON library we use?)